+++ Zolian.Server.Base/Models/ItemPending.cs
@@ -2,14 +2,14 @@
 
 public class PendingSell
 {
-    public uint ID { get; init; }
+    public long ID { get; init; }
     public string Name { get; init; }
     public ushort Quantity { get; set; }
 }
 
 public class PendingBuy
 {
-    public uint ID { get; init; }
+    public long ID { get; init; }
     public string Name { get; init; }
     public int Offer { get; init; }
     public ushort Quantity { get; set; }
